All is ready
