+++ src/test/java/com/pewpew/pewpew/rest/ErrorTest.java
@@ -2,7 +2,6 @@ package com.pewpew.pewpew.rest;
 
 import com.pewpew.pewpew.main.*;
 import com.pewpew.pewpew.model.User;
-import org.glassfish.hk2.utilities.binding.AbstractBinder;
 import org.glassfish.jersey.client.ClientConfig;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.test.JerseyTest;
@@ -149,18 +148,5 @@ public class ErrorTest extends JerseyTest {
         String id = "56f94e8a67bc2e7632974677";
         Response deleteJson = target("user").path(id).request().delete();
         assertEquals(deleteJson.getStatus(), Response.Status.FORBIDDEN.getStatusCode());
-    }
-
-    private static class MyAbstractBinder extends AbstractBinder {
-        private final Context context;
-
-        MyAbstractBinder(Context context) {
-            this.context = context;
-        }
-
-        @Override
-        protected void configure() {
-            bind(context);
-        }
     }
 }
